+++ prowler/CHANGELOG.md
@@ -2,7 +2,7 @@
 
 All notable changes to the **Prowler SDK** are documented in this file.
 
-## [5.27.0] (Prowler v5.27.0)
+## [5.27.0] (Prowler UNRELEASED)
 
 ### 🚀 Added